fix: scan foreign key entanglement on delete user fixed

+++ src/service/handlers/deleteUser.js
@@ -1,9 +1,21 @@
-const { queryDeleteUser } = require("../mysql.js");
+const { queryDeleteUser, queryDeleteScan } = require("../mysql.js");
 const authenticateUser = require("../authentication.js");
 
 const deleteUserByEmail = (email) => {
   return new Promise((resolve, reject) => {
     queryDeleteUser(email, (querySuccess, queryResults) => {
+      if (querySuccess) {
+        resolve(true);
+      } else {
+        reject(new Error("Query failed"));
+      }
+    });
+  });
+};
+
+const deleteScanById = (userId) => {
+  return new Promise((resolve, reject) => {
+    queryDeleteScan(userId, (querySuccess, queryResults) => {
       if (querySuccess) {
         resolve(true);
       } else {
@@ -24,17 +36,34 @@ const deleteUser = async (request, h) => {
     return response;
   }
 
-  const email = auth.decoded.email;
-  const deleteSuccess = await deleteUserByEmail(email);
+  const userId = auth.decoded.id;
+  const deleteRelatedScanSuccess = await deleteScanById(userId);
 
-  if (deleteSuccess) {
-    const response = h.response({
-      status: "success",
-      message: "successfully delete user",
-    });
-    response.code(202);
-    return response;
+  if (deleteRelatedScanSuccess) {
+
+    const email = auth.decoded.email;
+    const deleteSuccess = await deleteUserByEmail(email);
+
+    if (deleteSuccess) {
+      const response = h.response({
+        status: "success",
+        message: "successfully delete user",
+      });
+      response.code(202);
+      return response;
+    }
   }
+  // const email = auth.decoded.email;
+  // const deleteSuccess = await deleteUserByEmail(email);
+
+  // if (deleteSuccess) {
+  //   const response = h.response({
+  //     status: "success",
+  //     message: "successfully delete user",
+  //   });
+  //   response.code(202);
+  //   return response;
+  // }
   const response = h.response({
     status: "fail",
     message: "failed delete user",
